+++ src/distributions/index.js
@@ -5,4 +5,4 @@ export { Bernoulli } from './bernoulli.js';
 export { Beta } from './beta.js';
 export { Gamma } from './gamma.js';
 export { GaussianProcess } from './gp.js';
-export { RBF, Matern32, Matern52, Periodic, Linear } from './kernels.js';
+export { Kernel, RBF, Matern32, Matern52, Periodic, Linear } from './kernels.js';